Fixes `HiveTable.writeExecution` bugs.
It removes `MSCK REPAIR TABLE` statement since it is not needed. The
current use of it did not set the database and was broken.

It also changes the overwrite parts of the code to handle the case where
a table exists in the metastore but the underlying directory has been
deleted.

This fixes #426.