Fixes more partitioned overwrite bugs.
This fixes #382 again. This now changes partitioned files to have unique
names.

Previously, the file name were not unique which meant that sometimes old
and new filenames were the same and we would either not recognise that a
partition had been overwritten or delete a new file from a partition
because we thought it was an old file.